Add RVT reader
To support extensible autocomplete and error reporting a proper parser
needed to be created instead of relying on shlex. I decided on
having RVT to be a simple clojure inspired lisp that would look almost
identical to the current RVT syntax. The only difference is arguments
would now be lexed/scanned by the parser vs the tokens themselves.

The parser is not complete and will probably require some refactoring
but its a simple start and can support partial forms to support
autocomplete.